requirements.txt
absl-py==2.1.0
ai-edge-quantizer-nightly==0.0.1.dev20240718
ai-edge-torch==0.2.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloud-tpu-client==0.10
cloudpickle==3.0.0
decorator==5.1.1
dm-tree==0.1.8
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.9.0
gast==0.6.0
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
h5py==3.11.0
httplib2==0.22.0
idna==3.8
immutabledict==4.2.0
Jinja2==3.1.4
keras-nightly==3.5.0.dev2024090703
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
onnx==1.16.2
onnx-tf==1.10.0
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
six==1.16.0
soundfile==0.12.1
sox==1.5.0
sympy==1.13.2
tabulate==0.9.0
tb-nightly==2.18.0a20240906
tensorboard-data-server==0.7.2
tensorflow-addons==0.23.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.24.0
termcolor==2.4.0
tf_nightly==2.18.0.dev20240906
torch==2.4.1
torch-xla==2.4.0
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typeguard==2.13.3
typing_extensions==4.12.2
uritemplate==3.0.1
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
